requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
python-telegram-bot==13.7
google-cloud-storage==1.42.3
google-cloud-firestore==2.3.4
google-cloud-logging==2.6.0